+++ src/containers/Settings/Security/components/ChangePassword/changePassword.module.scss
@@ -1,0 +1,14 @@
+.input {
+  display: block;
+  margin-bottom: 14px;
+  input {
+    width: 450px;
+  }
+}
+
+.tip {
+  margin-top: 20px;
+  margin-bottom: 12px;
+  font-size: 12px;
+  color: #5f6368;
+}
