+++ src/containers/Settings/Security/components/TwoFactors/twoFactors.module.scss
@@ -1,0 +1,4 @@
+.arrowIcon {
+  width: 16px;
+  height: 16px;
+}
