+++ src/containers/Settings/components/SettingItemWrapper/settingItemWrapper.module.scss
@@ -26,18 +26,3 @@
     width: auto;
   }
 }
-
-.input {
-  display: block;
-  margin-bottom: 14px;
-  input {
-    width: 450px;
-  }
-}
-
-.tip {
-  margin-top: 20px;
-  margin-bottom: 4px;
-  font-size: 12px;
-  color: #5f6368;
-}
